realestate_analysis_backend/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
pandas==2.2.1
numpy==1.26.4
openpyxl==3.1.2
groq==0.6.0
python-dotenv==1.0.1


gunicorn==21.2.0
whitenoise==6.6.0


xlrd==2.0.1
xlsxwriter==3.1.9


django-cors-headers==4.3.1
